requirements.txt
asgiref==3.8.1
black==24.4.1
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.1
flake8==7.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
pycodestyle==2.12.0
pyflakes==3.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
